backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
numpy==1.26.4
openai==1.12.0
openai-whisper==20231117
